+++ app/src/main/assets/query_banks/ru/legal.json
@@ -434,7 +434,6 @@
   "путь к гражданству для нелегальных иммигрантов",
   "программа защиты молодых мигрантов DACA условия",
   "приостановление высылки из страны что это значит",
-  "защита от пыток при депортации что это",
   "как подготовиться к заседанию иммиграционного суда",
   "что такое добровольный выезд в иммиграционном праве",
   "отмена высылки как получить право",
